+++ examples/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "open-coroutine-examples"
+version = "0.1.0"
+edition = "2021"
+authors = ["zhangzicheng@apache.org"]
+description = "The examples for open-coroutine"
+repository = "https://github.com/acl-dev/open-coroutine/tree/dev/examples"
+keywords = ["open-coroutine", "example"]
+categories = ["os", "concurrency", "asynchronous"]
+license = "LGPL-3.0 OR Apache-2.0"
+
+[dependencies]
+open-coroutine = { version = "0.4.0", path = "../../open-coroutine-core-v2" }
+
+[[example]]
+name = "not_co"
+path = "not_co.rs"
+
+[[example]]
+name = "co_server"
+path = "co_server.rs"
+
+[[example]]
+name = "co_client"
+path = "co_client.rs"
+
+[[example]]
+name = "co"
+path = "co.rs"
