+++ src/tests/mod.rs
@@ -2,8 +2,6 @@ use super::*;
 
 mod sleep;
 
-mod socket;
-
 #[test]
 fn test_link() {
     init();
